Revert "Update UI, status bar style for List Survey"
