+++ Survey/Models/Survey.swift
@@ -11,11 +11,4 @@ struct Survey: Codable {
     var title: String?
     var description: String?
     var coverImageUrl: String?
-    
-    var highResolutionCoverImageUrl: String? {
-        guard let coverImageUrl = coverImageUrl else {
-            return nil
-        }
-        return "\(coverImageUrl)l"
-    }
 }
