+++ Survey/Modules/Survey/SurveyList/SurveyListViewController.swift
@@ -16,15 +16,8 @@ class SurveyListViewController: ViewController {
     private let onRefresh = PublishSubject<Void>()
     private let onTakeSurvey = PublishSubject<Survey>()
     private let onLoadMore = PublishSubject<Void>()
-    private let onPageChanged = BehaviorSubject<Int>(value: 0)
-    private var barStyle: UIStatusBarStyle?
-
-    private lazy var dateView = DateView()
     
-    private lazy var xxx: UIImageView = {
-        let imageView = UIImageView()
-        return imageView
-    }()
+    private lazy var dateView = DateView()
     
     private lazy var userImageView: UIImageView = {
         let imageView = UIImageView(asset: .userPicture)
@@ -79,10 +72,6 @@ class SurveyListViewController: ViewController {
         pageControl.hidesForSinglePage = true
         pageControl.isSkeletonable = true
         pageControl.skeletonCornerRadius = 8
-        if #available(iOS 14.0, *) {
-          pageControl.backgroundStyle = .minimal
-          pageControl.allowsContinuousInteraction = false
-        }
         pageControl.rx.controlEvent(.valueChanged)
             .subscribe(onNext: { [weak self] in
                 guard let self = self else { return }
@@ -109,7 +98,6 @@ class SurveyListViewController: ViewController {
             pageControl
         ])
         view.isSkeletonable = true
-        barStyle = preferredStatusBarStyle
         collectionView.snp.makeConstraints { make in
             make.edges.equalToSuperview()
         }
@@ -130,15 +118,8 @@ class SurveyListViewController: ViewController {
         }
         
         pageControl.snp.makeConstraints { make in
-            make.left.equalToSuperview().offset(20)
-            make.bottom.equalToSuperview().offset(-170)
-        }
-        
-        view.addSubview(xxx)
-        xxx.snp.makeConstraints { make in
-            make.centerY.equalToSuperview()
-            make.left.right.equalToSuperview()
-            make.height.equalTo(UIApplication.shared.statusBarFrame.height)
+            make.left.equalToSuperview()
+            make.bottom.equalToSuperview().offset(-180)
         }
     }
     
@@ -226,25 +207,6 @@ class SurveyListViewController: ViewController {
             }
             self?.view.hideSkeleton(reloadDataAfter: false, transition: .crossDissolve(0.25))
         }).disposed(by: rx.disposeBag)
-            
-        Observable.combineLatest(
-            output.surveys.asObservable(),
-            Observable.merge(
-                Observable.just(()),
-                collectionView.rx.didEndScrollingAnimation.asObservable(),
-                collectionView.rx.didEndDecelerating.asObservable()
-            )
-        )
-        .subscribe(
-            onNext: { [weak self] (surveys, _) in
-                guard let self = self else { return }
-                let page = Int(self.collectionView.contentOffset.x/UIScreen.main.bounds.width)
-                guard page < surveys.count, let url = surveys[page]?.highResolutionCoverImageUrl else { return }
-                print(url)
-                print(surveys.count)
-                print( "\(page)\n")
-                self.updateStatusBarByImage(url)
-            }).disposed(by: rx.disposeBag)
     }
     
     @objc func refresh() {
@@ -268,40 +230,5 @@ class SurveyListViewController: ViewController {
     
     override var defaultLoadingAnimation: Bool {
         return false
-    }
-    
-    override var preferredStatusBarStyle: UIStatusBarStyle {
-        return barStyle ?? super.preferredStatusBarStyle
-    }
-    
-    private func updateStatusBarByImage(_ urlString: String) {
-        guard let url = URL(string: urlString) else { return }
-        KingfisherManager.shared.retrieveImage(with: url) { [weak self] response in
-            guard let self = self else { return }
-            switch response {
-            case .success:
-                let image = self.view.asImage(
-                    bounds: CGRect(
-                        x: 0,
-                        y: 0,
-                        width: self.collectionView.frame.width,
-                        height:  UIApplication.shared.statusBarFrame.height
-                    )
-                )
-                let islight = image.isLight ?? false
-                self.updateStatusBar(light: !islight)
-            default:
-                self.updateStatusBar(light: false)
-            }
-        }
-    }
-    
-    private func updateStatusBar(light: Bool) {
-        if #available(iOS 13.0, *) {
-            self.barStyle = light ? .lightContent : .darkContent
-        } else {
-            self.barStyle = light ? .lightContent : .default
-        }
-        self.setNeedsStatusBarAppearanceUpdate()
     }
 }
